Restore files to their original or similar bitrates
Several of the music files here were aggressively re-encoded by me many
years ago with the intention of reducing download sizes. I greatly
regret this decision since it is fairly destructive and the right ears
paired with the right audio equipment can easily distinguish the
downgrade in quality resulting from this.

This commit restores some files to their original Ogg Vorbis versions
from back in the day with higher bitrate, or to Ogg Vorbis conversions
from MP3 with a higher bitrate; the sources for this are either
Pentarctagon's UMC Music Book add-ons (1.18 versions) or my own backups
of files dating all the way back to 2006 in some cases (Orc Theme and
gameplay06.ogg in particular).

Most importantly, this makes Lonely Snowfall NOT be a crappy 96 kbps
reencoding -- what the hell was I thinking?

NOTE:

The ambuscade-begin/loop.ogg files included in this commit are new
edits of identical length I just made based on the ambuscade.ogg
original; the latter I am *not* entirely sure if it really is the file
West originally provided to me for the specific purpose of using it
in my campaigns without the production outro that can be found in the
more common UMC Music Book version. I'm afraid that in this case I'm
just appealing to the fact that it has a higher bitrate, even if it's
not as high as the regular version. All I know is that I found it in
a folder named "AtS-music-precheckout" or so.